requirements.txt
# Core data processing libraries
numpy>=1.21.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.3.0
scikit-image>=0.18.0

# Video processing
moviepy>=1.0.3
imageio>=2.9.0

# Text processing
nltk>=3.6.0
spacy>=3.4.0
transformers>=4.20.0

# Machine Learning
scikit-learn>=1.0.0
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.4.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0